% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalTable1.R
\name{comparativeCharacterization}
\alias{comparativeCharacterization}
\title{Comparative Characterization}
\usage{
comparativeCharacterization(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema = NULL,
  outputFolder,
  minCellCount = 0,
  targetCohortId,
  comparatorCohortId = NULL,
  outcomeCohortIds = NULL,
  tableSpecification = setTableSpecification(),
  sampleSize = NULL,
  output = "one column",
  percentDigits = 1,
  valueDigits = 1,
  stdDiffDigits = 2,
  studyPopulationSetting = NULL,
  fileName = NULL
)
}
\description{
Comparative Characterization
}
\details{
This function create the comparative table 1 for target and comparator
}
