% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalTable1.R
\name{setTableSpecification}
\alias{setTableSpecification}
\title{Set the setting for the table 1}
\usage{
setTableSpecification(
  useDemographicsGender = FALSE,
  useDemographicsAge = FALSE,
  useDemographicsAgeGroup = FALSE,
  useDemographicsRace = FALSE,
  useDemographicsEthnicity = FALSE,
  useDemographicsIndexYear = FALSE,
  useDemographicsIndexMonth = FALSE,
  useDemographicsPriorObservationTime = FALSE,
  useDemographicsPostObservationTime = FALSE,
  useDemographicsTimeInCohort = FALSE,
  useDemographicsIndexYearMonth = FALSE,
  conceptIdsConditionOccurrenceAnyTimePrior = c(),
  conceptIdsConditionOccurrenceLongTerm = c(),
  conceptIdsConditionOccurrenceMediumTerm = c(),
  conceptIdsConditionOccurrenceShortTerm = c(),
  conceptIdsConditionOccurrencePrimaryInpatientAnyTimePrior = c(),
  conceptIdsConditionOccurrencePrimaryInpatientLongTerm = c(),
  conceptIdsConditionOccurrencePrimaryInpatientMediumTerm = c(),
  conceptIdsConditionOccurrencePrimaryInpatientShortTerm = c(),
  conceptIdsConditionEraAnyTimePrior = c(),
  conceptIdsConditionEraLongTerm = c(),
  conceptIdsConditionEraMediumTerm = c(),
  conceptIdsConditionEraShortTerm = c(),
  conceptIdsConditionEraOverlapping = c(),
  conceptIdsConditionEraStartLongTerm = c(),
  conceptIdsConditionEraStartMediumTerm = c(),
  conceptIdsConditionEraStartShortTerm = c(),
  conceptIdsConditionGroupEraAnyTimePrior = c(),
  conceptIdsConditionGroupEraLongTerm = c(),
  conceptIdsConditionGroupEraMediumTerm = c(),
  conceptIdsConditionGroupEraShortTerm = c(),
  conceptIdsConditionGroupEraOverlapping = c(),
  conceptIdsConditionGroupEraStartLongTerm = c(),
  conceptIdsConditionGroupEraStartMediumTerm = c(),
  conceptIdsConditionGroupEraStartShortTerm = c(),
  conceptIdsDrugExposureAnyTimePrior = c(),
  conceptIdsDrugExposureLongTerm = c(),
  conceptIdsDrugExposureMediumTerm = c(),
  conceptIdsDrugExposureShortTerm = c(),
  conceptIdsDrugEraAnyTimePrior = c(),
  conceptIdsDrugEraLongTerm = c(),
  conceptIdsDrugEraMediumTerm = c(),
  conceptIdsDrugEraShortTerm = c(),
  conceptIdsDrugEraOverlapping = c(),
  conceptIdsDrugEraStartLongTerm = c(),
  conceptIdsDrugEraStartMediumTerm = c(),
  conceptIdsDrugEraStartShortTerm = c(),
  conceptIdsDrugGroupEraAnyTimePrior = c(),
  conceptIdsDrugGroupEraLongTerm = c(),
  conceptIdsDrugGroupEraMediumTerm = c(),
  conceptIdsDrugGroupEraShortTerm = c(),
  conceptIdsDrugGroupEraOverlapping = c(),
  conceptIdsDrugGroupEraStartLongTerm = c(),
  conceptIdsDrugGroupEraStartMediumTerm = c(),
  conceptIdsDrugGroupEraStartShortTerm = c(),
  conceptIdsProcedureOccurrenceAnyTimePrior = c(),
  conceptIdsProcedureOccurrenceLongTerm = c(),
  conceptIdsProcedureOccurrenceMediumTerm = c(),
  conceptIdsProcedureOccurrenceShortTerm = c(),
  conceptIdsDeviceExposureAnyTimePrior = c(),
  conceptIdsDeviceExposureLongTerm = c(),
  conceptIdsDeviceExposureMediumTerm = c(),
  conceptIdsDeviceExposureShortTerm = c(),
  conceptIdsMeasurementAnyTimePrior = c(),
  conceptIdsMeasurementLongTerm = c(),
  conceptIdsMeasurementMediumTerm = c(),
  conceptIdsMeasurementShortTerm = c(),
  conceptIdsMeasurementValueAnyTimePrior = c(),
  conceptIdsMeasurementValueLongTerm = c(),
  conceptIdsMeasurementValueMediumTerm = c(),
  conceptIdsMeasurementValueShortTerm = c(),
  conceptIdsMeasurementRangeGroupAnyTimePrior = c(),
  conceptIdsMeasurementRangeGroupLongTerm = c(),
  conceptIdsMeasurementRangeGroupMediumTerm = c(),
  conceptIdsMeasurementRangeGroupShortTerm = c(),
  conceptIdsObservationAnyTimePrior = c(),
  conceptIdsObservationLongTerm = c(),
  conceptIdsObservationMediumTerm = c(),
  conceptIdsObservationShortTerm = c(),
  useCharlsonIndex = FALSE,
  useDcsi = FALSE,
  useChads2 = FALSE,
  useChads2Vasc = FALSE,
  useHfrs = FALSE,
  useDistinctConditionCountLongTerm = FALSE,
  useDistinctConditionCountMediumTerm = FALSE,
  useDistinctConditionCountShortTerm = FALSE,
  useDistinctIngredientCountLongTerm = FALSE,
  useDistinctIngredientCountMediumTerm = FALSE,
  useDistinctIngredientCountShortTerm = FALSE,
  useDistinctProcedureCountLongTerm = FALSE,
  useDistinctProcedureCountMediumTerm = FALSE,
  useDistinctProcedureCountShortTerm = FALSE,
  useDistinctMeasurementCountLongTerm = FALSE,
  useDistinctMeasurementCountMediumTerm = FALSE,
  useDistinctMeasurementCountShortTerm = FALSE,
  useDistinctObservationCountLongTerm = FALSE,
  useDistinctObservationCountMediumTerm = FALSE,
  useDistinctObservationCountShortTerm = FALSE,
  useVisitCountLongTerm = FALSE,
  useVisitCountMediumTerm = FALSE,
  useVisitCountShortTerm = FALSE,
  useVisitConceptCountLongTerm = FALSE,
  useVisitConceptCountMediumTerm = FALSE,
  useVisitConceptCountShortTerm = FALSE,
  longTermStartDays = -365,
  mediumTermStartDays = -180,
  shortTermStartDays = -30,
  endDays = 0
)
}
\arguments{
\item{useDemographicsGender}{= FALSE}

\item{endDays}{}
}
\value{
An object of type \code{tableSpecification}, to be used in other functions.
}
\description{
Set the setting for the table 1
}
\details{
This function create the setting for the table 1.
}
